Rename function directory
