+++ app/scripts.babel/functions/formatDate.js
